tftp: add support of blksize option to client
The TFTP protocol uses a default block size of 512 bytes. This value is
sub-optimal for ethernet devices, which have a MTU (Maximum Transmission
Unit) of 1500 bytes. When taking into acount the overhead of the IP and
UDP layers, this leaves 1468 bytes for the TFTP payload.

This patch introduces a new function: tftp_client_set_blksize() which
may be used to change the block size from the default. It has to be
called after tftp_client_init() and before tftp_get(). If the server
does not support the option, the client will still accept to receive
512-byte blocks.

Signed-off-by: Jerome Forissier <jerome.forissier@linaro.org>
[For a fix in find_option()]
Co-developed-by: Heinrich Schuchardt <heinrich.schuchardt@canonical.com>
Link: https://source.denx.de/u-boot/u-boot/-/commit/27d7ccda94f
Link: https://source.denx.de/u-boot/u-boot/-/commit/10917df17f2