altcp_tls: set errno to EPERM when handshake fails
When a TLS client connects to a server and the handshake fails due to
certificate validation, no specific error code is returned to the
caller. The TCP connection is simply closed. This is not very
user-friendly. Instead the client application would probably want to
notify the user of the security issue. For that purpose, set errno to
EPERM to indicate that a validation error occured.

This change is upstreamed from U-Boot v2025.07-rc2-26-g9349fc2e9c7 [1].

Link: https://source.denx.de/u-boot/u-boot/-/commit/9349fc2e9c76d042f424edfa69cf421225bf0fcc [1]
Signed-off-by: Jerome Forissier <jerome.forissier@linaro.org>