contrib: apps: tcpecho_altcp: an altcp variant of tcpecho
The altcp is a wonderful mechanism which wraps the raw api and provides the
ability to use TLS or other layers. It is also automatically reverted to the
raw api  with the LWIP_ALTCP is disabled.

Having an example of using altcp will allow quick starting point for people
who would like to leverage layers.

Signed-off-by: Alon Bar-Lev <alon.barlev@gmail.com>